Move src to new security_settings.rs
